added linux update scripts

+++ scripts_staging/linux_os_update.sh
@@ -1,18 +1,43 @@
 #!/bin/bash
 
-#Update script to run for most common linux distros
+# Synopsis: This script automates the process of updating software packages across multiple Linux distributions.
+# It checks for the available package manager (dnf, yum, apt, pacman, or zypper) and executes the appropriate commands to update the system.
+# Users can optionally allow the script to automatically reboot the system after updates by passing the --autoreboot flag.
+#
+# Usage:
+# Update with automatic reboot --autoreboot
+#
+# Note: The script is designed to be flexible, catering both to interactive use cases and automated workflows.
 
-if [[ `which yum` ]]; then
+AUTO_REBOOT=0
+
+# Check for --autoreboot flag
+for arg in "$@"; do
+    if [[ $arg == "--autoreboot" ]]; then
+        AUTO_REBOOT=1
+    fi
+done
+
+# Update system based on package manager availability
+if command -v dnf &> /dev/null; then
+   dnf -y update
+elif command -v yum &> /dev/null; then
    yum -y update
-elif [[ `which apt` ]]; then
-   apt-get -y update
-   apt-get -y upgrade
-elif [[ `which pacman` ]]; then
+elif command -v apt &> /dev/null; then
+   apt-get -y update && apt-get -y upgrade
+elif command -v pacman &> /dev/null; then
    pacman -Syu
-elif [[ `which zypper` ]]; then
+elif command -v zypper &> /dev/null; then
    zypper update
 else
-   echo "Unknown Platform"
+   echo "Package manager not detected. Please update your system manually."
+   exit 1
 fi
 
-sleep 10 && reboot &
+# Handle auto-reboot
+if [ $AUTO_REBOOT -eq 1 ]; then
+    echo "Rebooting in 10 seconds..."
+    sleep 10 && reboot &
+else
+   echo "Updates done, please reboot"
+fi
